backend/requirements.txt
# Django core
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3

# Database
psycopg2-binary

# Background tasks
celery[redis]

# Payment processing
stripe

# Environment variables
django-environ
python-dotenv

# Image processing
Pillow>=10.0

# HTTP requests (for AI API calls)
requests>=2.31

# Static files
whitenoise>=6.6

# Production server
gunicorn>=21.0
